requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
bootstrap4==0.1.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contextlib2==21.6.0
coverage==7.6.1
DateTime==5.5
dj-database-url==2.2.0
Django==4.2.15
django-cors-headers==4.4.0
django-decouple==2.1
django-extensions==3.2.3
django-filter==24.3
django-safedelete==1.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.2
execnet==2.1.1
gunicorn==23.0.0
idna==3.8
inflection==0.5.1
iniconfig==2.0.0
kombu==5.4.0
mock==5.1.0
packaging==24.1
path==17.0.0
path.py==12.5.0
pillow==10.4.0
pluggy==1.5.0
prompt-toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.2
pytest-shutil==1.7.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
SafeDelete==0.0.2
six==1.16.0
sqlparse==0.5.1
termcolor==2.4.0
tomli==2.0.1
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
zope.interface==7.0.3
